Implement `LoadingScene` with resource management and initialization
